% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runCrossValidation.R
\name{calculatePerformanceCross}
\alias{calculatePerformanceCross}
\title{calculatePerformanceCross}
\usage{
calculatePerformanceCross(db, conf, N = 10, m = 3, keepMatrices = NA)
}
\arguments{
\item{db}{Database of timeseries.}

\item{conf}{Configuration for DTW algorithm as a data.frame row.}

\item{N}{Number of rounds, with one different reference set each round.}

\item{m}{Number of test sets per round.}

\item{keepMatrices}{Save intermediate cross DTW matrices in files. Defaults to NA.}
}
\value{
Average matching rate after N rounds.
}
\description{
Calculates aprox. matching rate for given a database and configuration
}
\examples{
calculatePerformanceCross(db, conf, keepMatrices="dbs/crossMatrices/")
}
