% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createTraceDatabase.R
\name{createTraceDatabase}
\alias{createTraceDatabase}
\title{createTraceDatabase}
\usage{
createTraceDatabase(inPath, outPath, maxLength = 4000)
}
\arguments{
\item{inPath}{Path with raw data.}

\item{maxLength}{Maximum duration length of vectors. Defaults to 4000 milliseconds.}

\item{output}{Output file for writing.}
}
\value{
list with vectors of delays.
}
\description{
Lists all files in 'inPath' and generates a list of data.frames (timestamp/delay)
}
\examples{
createTraceDatabase("data/", "traces/")
}

