% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawHeatmap.R
\name{drawHeatmap}
\alias{drawHeatmap}
\title{Calculates and plots a heatmap of the "confusion matrix" M for the given subset of pages.
M[i,j] is the probability that a 1-NN classifier classifies a random query from
class i as belonging to j, given our dataset.}
\usage{
drawHeatmap(filesPath, webpages = "all", showPercentages = F, P = 5,
  windowType = "sakoechiba", windowSize = 100, maxLength = 1000,
  sampFun = "sum", stepPattern = "symmetric1", pathOut,
  matrixFileName = "confusionMatrix", heatmapFileName = "heatmap")
}
\arguments{
\item{filesPath}{Path containing the crossDistance matrices.}

\item{webpages}{List of the pages you want to restrict the classifier to.}

\item{showPercentages}{Logical. TRUE displays the value of M[i,j] in the
corresponding cell. Defaults to FALSE.}

\item{P}{Timeseries averaging parameter}

\item{pathOut}{Path to save the confusion matrix and heatmap image.}

\item{matrixFileName}{R object file containing the confusion matrix.}

\item{heatmapFileName}{pdf file to save the heatmap to.}

\item{weight}{Either 'none' (default) for simple majority vote, 'inverse' for
1/distance weighting.}
}
\description{
Calculates and plots a heatmap of the "confusion matrix" M for the given subset of pages.
M[i,j] is the probability that a 1-NN classifier classifies a random query from
class i as belonging to j, given our dataset.
}
