% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/executeTuning.R, R/runCrossValidation.R
\name{dtwDistParallel}
\alias{dtwDistParallel}
\alias{dtwDistParallel}
\title{dtwDistParallel}
\usage{
dtwDistParallel(mx, my, conf, tag, output = NA)

dtwDistParallel(mx, my, conf, tag, output = NA)
}
\arguments{
\item{mx}{List of timeseries.}

\item{my}{List of timeseries.}

\item{conf}{Parameters for 'dtw' function.}

\item{output}{Output folder to save the cross DTW distance matrix. Defaults to NA.}

\item{mx}{List of timeseries.}

\item{my}{List of timeseries.}

\item{conf}{Parameters for 'dtw' function.}

\item{output}{Output folder to save the cross DTW distance matrix. Defaults to NA.}
}
\value{
Returns the cross DTW distance matrix.

Returns the cross DTW distance matrix.
}
\description{
Calculates cross DTW distance matrix for X and Y (lists of timeseries)

Calculates cross DTW distance matrix for X and Y (lists of timeseries)
}
\examples{
dtwDistParallel()
dtwDistParallel()
}
