% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runCrossValidation.R
\name{executeCrossValidation}
\alias{executeCrossValidation}
\title{executeCrossValidation}
\usage{
executeCrossValidation(inPath, conf, N, m, logFile = NA)
}
\arguments{
\item{conf}{DTW parameters including step pattern and window type & size.}

\item{N}{Number of rounds in the cross validation.}

\item{m}{Number of test samples per round.}

\item{logFile}{Name of log file, defaults to NA}
}
\value{
perf Performance vector averaged over all m*N classifications.
}
\description{
Executes N-fold cross validation with m tests per round for a fixed set of DTW parameters
}
