% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/executeTuning.R
\name{executeTuning}
\alias{executeTuning}
\title{generateTuningDatabases}
\usage{
executeTuning(inPath, logFile = NA)
}
\arguments{
\item{inPath}{Path with databases of timeseries.}

\item{N}{Number of rounds with 1 dataset as ref and 1 as test.}
}
\value{
Data frame average results.
}
\description{
Creates databases of timeseries for all combinations of parameters
}
\examples{
executeTuning("dbs/")
}
