% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateTuningDatabases.R
\name{generateTuningDatabases}
\alias{generateTuningDatabases}
\title{generateTuningDatabases}
\usage{
generateTuningDatabases(inPath, outPath, RL, RP, RF)
}
\arguments{
\item{inPath}{Path with raw data}

\item{outPath}{Path for the output dbs.}

\item{RL}{Range of maximum length for the traces.}

\item{RP}{Range of P values.}

\item{RF}{Range of sampling functions.}
}
\value{
None
}
\description{
Creates databases of timeseries for all combinations of parameters
}
\examples{
generateTuningDatabases("data/", "dbs/", RL=c(4000,2000,1000), RP=c(50,20,10,5), RF=c("sum"))
}

