% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createRawDatabase.R
\name{loadRawDatabase}
\alias{loadRawDatabase}
\title{loadRawDatabase}
\usage{
loadRawDatabase(file)
}
\arguments{
\item{file}{Path with db.dat.}
}
\value{
List of pages with a vector of timestamps per page.
}
\description{
Load DB from 'file'
}
\examples{
loadRawDatabase("data/test.dat")
}

