% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traceDatabaseToTimeseries.R
\name{loadTimeseriesDatabase}
\alias{loadTimeseriesDatabase}
\title{loadTimeseriesDatabase}
\usage{
loadTimeseriesDatabase(file)
}
\arguments{
\item{file}{Path with db.dat.}
}
\value{
List of timeseries.
}
\description{
Load DB from 'file'
}
\examples{
loadTimeseriesDatabase("data/test.dat")
}

