% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createTraceDatabase.R
\name{loadTraceDatabase}
\alias{loadTraceDatabase}
\title{loadTraceDatabase}
\usage{
loadTraceDatabase(path)
}
\arguments{
\item{file}{Path with db.dat.}
}
\value{
List of data.frames (timestamp/delay).
}
\description{
Load trace DB from path
}
\examples{
loadTraceDatabase("traces/traces_5_4000.dat")
}

