% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paintMultiTimeseries.R
\name{paintMultiTimeseries}
\alias{paintMultiTimeseries}
\title{paintMultiTimeseries}
\usage{
paintMultiTimeseries(tss)
}
\arguments{
\item{v}{List of timeseries.}
}
\value{
None.
}
\description{
Draws a line plot from several timeseries.
}
\examples{
paintMultiTimseries(list(ts1, ts2, ts3))
}

