% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paintMultiTraces.R
\name{paintMultiTraces}
\alias{paintMultiTraces}
\title{paintMultiTraces}
\usage{
paintMultiTraces(traces)
}
\arguments{
\item{v}{List of traces (data.frames).}
}
\value{
None.
}
\description{
Draws a line plot from several traces.
}
\examples{
paintMultiTraces(list(tr1, tr2, tr3))
}
