% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paintTracesList.R
\name{paintTracesList}
\alias{paintTracesList}
\title{paintTracesList}
\usage{
paintTracesList(v)
}
\arguments{
\item{v}{List of traces (data.frames).}
}
\value{
Graphic to plot.
}
\description{
Generates a graphic with several traces using grid.arrange
}
\examples{
paintTracesList(list(tr1, tr2, tr3))
}
