% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/executeTuning.R, R/runCrossValidation.R
\name{parseDatabaseFilename}
\alias{parseDatabaseFilename}
\alias{parseDatabaseFilename}
\title{parseDatabaseFilename}
\usage{
parseDatabaseFilename(name)

parseDatabaseFilename(name)
}
\arguments{
\item{name}{Filename.}

\item{name}{Filename.}
}
\value{
Named list of parameters.

Named list of parameters.
}
\description{
Extract parameters from db filename

Extract parameters from db filename
}
\examples{
parseDatabaseFilename("ts_1000_20_sum.db")
parseDatabaseFilename("ts_1000_20_sum.db")
}
