% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseTrace.R
\name{parseTrace}
\alias{parseTrace}
\title{parseTrace}
\usage{
parseTrace(trace)
}
\arguments{
\item{trace}{Vector of timestamps}
}
\value{
data.frame of delays and timestamps.
}
\description{
Converts a vector of timestamps 'trace' into a data.frame (timestamp/delay)
}
\examples{
parseTrace("db$example_0.json")
}
