% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/executeTuning.R, R/runCrossValidation.R
\name{printConfigData}
\alias{printConfigData}
\alias{printConfigData}
\title{printConfigData}
\usage{
printConfigData(conf)

printConfigData(conf)
}
\arguments{
\item{conf}{Config data.frame.}

\item{conf}{Config data.frame.}
}
\value{
String.

String.
}
\description{
Returns formatted config data string

Returns formatted config data string
}
