% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/executeTuning.R
\name{recognitionRate}
\alias{recognitionRate}
\title{recognitionRate}
\usage{
recognitionRate(M)
}
\arguments{
\item{M}{Cross distance matrix.}

\item{tolerance}{Tolerance. Defaults to 1.}
}
\value{
Recognition rate.
}
\description{
Given a cross distance matrix returns recognition rate
}

