% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runCrossValidation.R
\name{runCrossValidation}
\alias{runCrossValidation}
\title{runCrossValidation}
\usage{
runCrossValidation(inPath, windowTypes = c("sakoechiba", "slantedband",
  "itakura"), stepPatterns = c("symmetric1", "symmetric2", "asymmetric"),
  windowSizes = c(1, 5, 10, 20, 50, 100), P = c(1, 2, 5, 10, 20, 50),
  maxLength = c(1000, 2000, 3000, 4000), sampFun = c("sum", "max"), N, m,
  logFile = NA)
}
\arguments{
\item{inPath}{Path containing timeseries databases.}

\item{windowTypes}{Window types.}

\item{stepPatterns}{Step patterns.}

\item{windowSizes}{Window sizes.}

\item{P}{Time averaging parameter.}

\item{maxLength}{Maximum trace length}

\item{sampFun}{Sampling function.}

\item{N}{Number of cross validation rounds - one per reference sample.}

\item{m}{Number of test samples per round.}

\item{logFile}{Log file, defaults to NA.}
}
\value{
results Summary table with performance for each combination of
parameters.
}
\description{
Runs an N-fold cross validation with m test samples per round for all possible combinations of parameters
}
\examples{
runCrossValidation('dbs/')
}
