% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traceDatabaseToTimeseries.R
\name{traceDatabaseToTimeseries}
\alias{traceDatabaseToTimeseries}
\title{traceDatabaseToTimeseries}
\usage{
traceDatabaseToTimeseries(traceDatabase, P = 20, samplingFun = sum,
  output = NA)
}
\arguments{
\item{traceDatabase}{List of files with data.frames (timestamp/delay).}

\item{P}{Interval duration in milliseconds. Defaults to 20 ms.}

\item{samplingFun}{Sampling function. Defaults to 'sum'.}

\item{output}{Output file for writing.}
}
\value{
list of timeseries
}
\description{
Converts a list of data.frames (timestamp/delay) into a list of timeseries
}
\examples{
traceDatabaseToTimeseries(db, P=20, samplingFun=sum) 
}

