% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/executeTuning.R, R/runCrossValidation.R
\name{tracesMatch}
\alias{tracesMatch}
\alias{tracesMatch}
\title{tracesMatch}
\usage{
tracesMatch(query, response, tolerance = 1)

tracesMatch(query, response, tolerance = 1)
}
\arguments{
\item{tolerance}{Tolerance. Defaults to 1.}

\item{a}{Query's trace name}

\item{b}{List of names ordered by proximity.}

\item{a}{Query's trace name}

\item{b}{List of names ordered by proximity.}

\item{tolerance}{Tolerance. Defaults to 1.}
}
\value{
TRUE if query's name is in "tolerance" firsts sublist.

TRUE if query's name is in "tolerance" firsts sublist.
}
\description{
Returns true if trace's names are from same page

Returns true if trace's names are from same page
}
