% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traceToTimeseries.R
\name{traceToTimeseries}
\alias{traceToTimeseries}
\title{traceToTimeseries}
\usage{
traceToTimeseries(trace, P = 20, samplingFun = sum)
}
\arguments{
\item{trace}{Trace to convert.}

\item{P}{Interval duration in milliseconds. Defaults to 20 ms.}

\item{samplingFun}{Sampling function. Defaults to 'sum'.}
}
\value{
ts() object.
}
\description{
This function converts a raw Trace of time delays into a Timeserie sampling intervals of duration P
}
\examples{
traceToTimseries(trace, P=20)
}
