% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/executeTuning.R
\name{calculatePerformance}
\alias{calculatePerformance}
\title{calculatePerformance}
\usage{
calculatePerformance(db, conf, N = 3, keepMatrices = NA)
}
\arguments{
\item{db}{Database of timeseries.}

\item{conf}{Configuration for DTW algorithm as a data.frame row.}

\item{N}{Number of rounds with 1 set as ref and 1 as test. Defaults to 3.}

\item{keepMatrices}{Save intermediate cross DTW matrices in files. Defaults to NA.}
}
\value{
Average matching rate after N rounds.
}
\description{
Calculates aprox. matching rate for given a database and configuration
}
\examples{
calculatePerformance(db, conf, keepMatrices="dbs/crossMatrices/")
}

