% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/executeTuning.R
\name{printConfigData}
\alias{printConfigData}
\title{printConfigData}
\usage{
printConfigData(conf)
}
\arguments{
\item{conf}{Config data.frame.}
}
\value{
String.
}
\description{
Returns formatted config data string
}

